Decided to forego the coffeescript.
Apparently its use is discouraged, and I didn't like the JavaScript the
compiler output. Not only that, it adds unneccessary complexity to the
build process, and it didn't even make the code that much more concise.

This commit is here to merge changes to the actual javascript that I
made during the coffee fork.